Reorganize inference modules into callbacks as private utilities
