--- tests/inference/__init__.py
@@ -1,5 +1,0 @@
-# Copyright (c) TorchGeo Contributors
-# All rights reserved.
-# Licensed under the MIT License.
-
-"""Tests for inference utilities."""
