+++ torchgeo/callbacks/_writer.py
