--- torchgeo/inference/__init__.py
@@ -1,9 +1,0 @@
-# Copyright (c) TorchGeo Contributors
-# All rights reserved.
-# Licensed under the MIT License.
-
-"""TorchGeo inference utilities."""
-
-from torchgeo.inference.blending import get_blend_mask, weighted_merge
-
-__all__ = ['get_blend_mask', 'weighted_merge']
